Rename SUGroup -> SpecialUnitary
