+++ geometric_kernels/spaces/su.py
@@ -1,6 +1,6 @@
 """
-This module provides the :class:`SUGroup` space and the representation of
-its spectrum, the :class:`SUEigenfunctions` class.
+This module provides the :class:`SpecialUnitary` space and the representation
+of its spectrum, the :class:`SUEigenfunctions` class.
 """
 import itertools
 import json
@@ -129,7 +129,7 @@ class SUCharacter(LieGroupCharacter):
         return char_val
 
 
-class SUGroup(MatrixLieGroup):
+class SpecialUnitary(MatrixLieGroup):
     r"""
     The GeometricKernels space representing the special unitary group
     :math:`SU(n)` consisting of n by n complex unitary matrices with unit
