+++ tests/spaces/test_lie_groups.py
@@ -8,8 +8,8 @@ from opt_einsum import contract as einsum
 
 from geometric_kernels.kernels import MaternKarhunenLoeveKernel
 from geometric_kernels.kernels.feature_maps import RandomPhaseFeatureMapCompact
-from geometric_kernels.spaces.so import SpecialOrthogonal
-from geometric_kernels.spaces.su import SUGroup
+from geometric_kernels.spaces import SpecialOrthogonal
+from geometric_kernels.spaces import SpecialUnitary
 
 
 @pytest.fixture(name="group_cls", params=["so", "su"])
@@ -17,7 +17,7 @@ def _group_cls(request):
     if request.param == "so":
         return SpecialOrthogonal
     elif request.param == "su":
-        return SUGroup
+        return SpecialUnitary
 
 
 @pytest.fixture(name="group", params=[3, 5])
@@ -35,7 +35,7 @@ def _group_and_eigf(group, request):
 def get_dtype(group):
     if isinstance(group, SpecialOrthogonal):
         return np.double
-    elif isinstance(group, SUGroup):
+    elif isinstance(group, SpecialUnitary):
         return np.cdouble
     else:
         raise ValueError()
